// src/generated/models/electionItemViewResponse.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ElectionItemViewResponse {
  electionItemId?: number;
  displayOrder?: number;
  title?: string;
  imageUrl?: string;
  voteCount?: number;
  voteRate?: number;
  selected?: boolean;
}